+++ miner/src/main/java/fr/raksrinana/channelpointsminer/miner/api/chat/irc/TwitchIrcChatClient.java
@@ -1,7 +1,6 @@
 package fr.raksrinana.channelpointsminer.miner.api.chat.irc;
 
 import fr.raksrinana.channelpointsminer.miner.api.chat.ITwitchChatClient;
-import fr.raksrinana.channelpointsminer.miner.api.chat.TwitchChatFactory;
 import fr.raksrinana.channelpointsminer.miner.api.passport.TwitchLogin;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
@@ -39,11 +38,11 @@ public class TwitchIrcChatClient implements ITwitchChatClient{
 		if(Objects.isNull(ircClient)){
 			log.info("Creating new Twitch IRC client");
 			
-			ircClient = TwitchChatFactory.createIrcClient(twitchLogin);
+			ircClient = TwitchIrcFactory.createIrcClient(twitchLogin);
 			ircClient.connect();
 			ircClient.setExceptionListener(e -> log.error("Error from irc", e));
 			
-			ircClient.getEventManager().registerEventListener(TwitchChatFactory.createIrcListener(twitchLogin.getUsername()));
+			ircClient.getEventManager().registerEventListener(TwitchIrcFactory.createIrcListener(twitchLogin.getUsername()));
 			
 			log.info("IRC Client created");
 		}
